common_time::duration

Struct Duration

pub struct Duration {
    value: i64,
    unit: TimeUnit,
}

Duration represents the elapsed time in either seconds, milliseconds, microseconds or nanoseconds.

Fields
value: i64unit: TimeUnit
Implementations

impl Duration

pub fn new(value: i64, unit: TimeUnit) -> Self

Create a new Duration with value and TimeUnit.

pub fn new_second(value: i64) -> Self

Create a new Duration in second.

pub fn new_millisecond(value: i64) -> Self

Create a new Duration in millisecond.

pub fn new_microsecond(value: i64) -> Self

Create a new Duration in microsecond.

pub fn new_nanosecond(value: i64) -> Self

Create a new Duration in nanosecond.

pub fn unit(&self) -> TimeUnit

Return the TimeUnit of current Duration.

pub fn value(&self) -> i64

Return the value of current Duration.

fn split(&self) -> (i64, u32)

Split a Duration into seconds part and nanoseconds part. Notice the seconds part of split result is always rounded down to floor.

pub fn to_std_duration(self) -> Duration

Convert to std::time::Duration.

pub fn negative(self) -> Self