Module common_time::interval

Structs

Interval

Interval Type represents a period of time. It is composed of months, days and nanoseconds. 3 kinds of interval are supported: year-month, day-time and month-day-nano, which will be stored in the following format. Interval data format: | months | days | nsecs | | 4bytes | 4bytes | 8bytes |

IntervalCompare 🔒

IntervalCompare is used to compare two intervals It makes interval into nanoseconds style.

IntervalFormat

https://www.postgresql.org/docs/current/datatype-datetime.html#DATATYPE-INTERVAL-OUTPUT support postgres format, iso8601 format and sql standard format

Enums

IntervalUnit

Constants

DAYS_PER_MONTH

NANOS_PER_DAY

NANOS_PER_HOUR

NANOS_PER_MICRO

NANOS_PER_MILLI

NANOS_PER_MONTH

NANOS_PER_SEC

Functions

get_time_part 🔒

get time part string

get_year_month 🔒

get year month string