Struct common_time::interval::Interval

pub struct Interval {
    months: i32,
    days: i32,
    nsecs: i64,
    unit: IntervalUnit,
}

Interval Type represents a period of time. It is composed of months, days and nanoseconds. 3 kinds of interval are supported: year-month, day-time and month-day-nano, which will be stored in the following format. Interval data format: | months | days | nsecs | | 4bytes | 4bytes | 8bytes |

Fields
months: i32days: i32nsecs: i64unit: IntervalUnit
Implementations

impl Interval

pub fn from_month_day_nano(months: i32, days: i32, nsecs: i64) -> Self

Creates a new interval from months, days and nanoseconds. Precision is nanosecond.

pub fn from_year_month(months: i32) -> Self

Creates a new interval from months.

pub fn from_day_time(days: i32, millis: i32) -> Self

Creates a new interval from days and milliseconds.

pub fn to_duration(&self) -> Result<Duration>

pub fn to_month_day_nano(&self) -> (i32, i32, i64)

Return a tuple(months, days, nanoseconds) from the interval.

pub fn to_nanosecond(&self) -> i128

Converts the interval to nanoseconds.

pub const MIN: Self = _

Smallest interval value.

pub const MAX: Self = _

Largest interval value.

pub fn justified_interval(&self) -> Self

Returns the justified interval. allows you to adjust the interval of 30-day as one month and the interval of 24-hour as one day

pub fn is_positive(&self) -> bool

Convert Interval to nanoseconds, to check whether Interval is positive

pub fn is_zero(&self) -> bool

is_zero

pub fn unit(&self) -> IntervalUnit

get unit

pub fn checked_mul_int<I>(&self, rhs: I) -> Option<Self>
where I: TryInto<i32>,

Multiple Interval by an integer with overflow check. Returns justified Interval, or None if overflow occurred.

pub fn to_iso8601_string(self) -> String

Convert Interval to ISO 8601 string

pub fn to_postgres_string(self) -> String

Convert Interval to postgres verbose string

pub fn to_sql_standard_string(self) -> String

Convert Interval to sql_standard string

pub fn from_i128(v: i128) -> Self

Interval Type and i128 IntervalUnit::MonthDayNano Convert v consists of months(i32) | days(i32) | nsecs(i64)

pub fn from_i64(v: i64) -> Self

Interval Type and i64 IntervalUnit::DayTime Convert v consists of days(i32) | milliseconds(i32)

pub fn from_i32(v: i32) -> Self

Interval Type and i32 IntervalUnit::YearMonth Convert v consists of months(i32)

pub fn to_i128(&self) -> i128

pub fn to_i64(&self) -> i64

pub fn to_i32(&self) -> i32