Type Alias common_time::range::RangeMillis

pub type RangeMillis = GenericRange<TimestampMillis>;

Time range in milliseconds.

Aliased Type

struct RangeMillis {
    start: Option<TimestampMillis>,
    end: Option<TimestampMillis>,
}

Fields
start: Option<TimestampMillis>end: Option<TimestampMillis>
Implementations

impl<T> GenericRange<T>where T: Copy + PartialOrd + Default,

pub fn and(&self, other: &GenericRange<T>) -> GenericRange<T>

Computes the AND’ed range with other.

pub fn or(&self, other: &GenericRange<T>) -> GenericRange<T>

Compute the OR’ed range of two ranges. Notice: this method does not compute the exact OR’ed operation for simplicity. For example, [1, 2) or’ed with [4, 5) will produce [1, 5) instead of [1, 2) ∪ [4, 5)

pub fn intersects(&self, target: &GenericRange<T>) -> bool

Checks if current range intersect with target.

pub fn empty() -> GenericRange<T>

Create an empty range.

fn from_optional(start: Option<T>, end: Option<T>) -> GenericRange<T>

Create GenericRange from optional start and end. If the present value of start >= the present value of end, it will return an empty range with the default value of T.

impl<T> GenericRange<T>

pub fn new<U: PartialOrd + Into<T>>(start: U, end: U) -> Option<GenericRange<T>>

Creates a new range that contains values in [start, end).

Returns None if start > end.

pub fn min_to_max() -> GenericRange<T>

Return a range containing all possible values.

pub fn start(&self) -> &Option<T>

Returns the lower bound of the range (inclusive).

pub fn end(&self) -> &Option<T>

Returns the upper bound of the range (exclusive).

pub fn contains<U: PartialOrd<T>>(&self, target: &U) -> bool

Returns true if timestamp is contained in the range.

impl<T: PartialOrd> GenericRange<T>

pub fn is_empty(&self) -> bool

Returns true if the range contains no timestamps.