common_time::timestamp

Struct Timestamp

pub struct Timestamp {
    value: i64,
    unit: TimeUnit,
}

Timestamp represents the value of units(seconds/milliseconds/microseconds/nanoseconds) elapsed since UNIX epoch. The valid value range of Timestamp depends on it’s unit (all in UTC timezone):

for TimeUnit::Second: [-262144-01-01 00:00:00, +262143-12-31 23:59:59]

for TimeUnit::Millisecond: [-262144-01-01 00:00:00.000, +262143-12-31 23:59:59.999]

for TimeUnit::Microsecond: [-262144-01-01 00:00:00.000000, +262143-12-31 23:59:59.999999]

for TimeUnit::Nanosecond: [1677-09-21 00:12:43.145224192, 2262-04-11 23:47:16.854775807]

Note:

For values out of range, you can still store these timestamps, but while performing arithmetic or formatting operations, it will return an error or just overflow.

Fields
value: i64unit: TimeUnit
Implementations

impl Timestamp

pub fn current_millis() -> Self

Creates current timestamp in millisecond.

pub fn current_time(unit: TimeUnit) -> Timestamp

Creates current timestamp in specific time unit.

pub fn sub_duration(&self, duration: Duration) -> Result<Self>

Subtracts a duration from timestamp.

Note

The result time unit remains unchanged even if duration has a different unit with self. For example, a timestamp with value 1 and time unit second, subtracted by 1 millisecond and the result is still 1 second.

pub fn add_duration(&self, duration: Duration) -> Result<Self>

Adds a duration to timestamp.

Note

The result time unit remains unchanged even if duration has a different unit with self. For example, a timestamp with value 1 and time unit second, subtracted by 1 millisecond and the result is still 1 second.

pub fn add_year_month(&self, interval: IntervalYearMonth) -> Option<Timestamp>

Adds given IntervalYearMonth to the current timestamp.

pub fn add_day_time(&self, interval: IntervalDayTime) -> Option<Timestamp>

Adds given IntervalDayTime to the current timestamp.

pub fn add_month_day_nano( &self, interval: IntervalMonthDayNano, ) -> Option<Timestamp>

Adds given IntervalMonthDayNano to the current timestamp.

pub fn sub_year_month(&self, interval: IntervalYearMonth) -> Option<Timestamp>

Subtracts given IntervalYearMonth to the current timestamp.

pub fn sub_day_time(&self, interval: IntervalDayTime) -> Option<Timestamp>

Subtracts given IntervalDayTime to the current timestamp.

pub fn sub_month_day_nano( &self, interval: IntervalMonthDayNano, ) -> Option<Timestamp>

Subtracts given IntervalMonthDayNano to the current timestamp.

pub fn sub(&self, rhs: &Self) -> Option<Duration>

Subtracts current timestamp with another timestamp, yielding a duration.

pub fn new(value: i64, unit: TimeUnit) -> Self

pub const fn new_second(value: i64) -> Self

pub const fn new_millisecond(value: i64) -> Self

pub const fn new_microsecond(value: i64) -> Self

pub const fn new_nanosecond(value: i64) -> Self

pub fn unit(&self) -> TimeUnit

pub fn value(&self) -> i64

pub fn convert_to(&self, unit: TimeUnit) -> Option<Timestamp>

Convert a timestamp to given time unit. Conversion from a timestamp with smaller unit to a larger unit may cause rounding error. Return None if conversion causes overflow.

pub fn convert_to_ceil(&self, unit: TimeUnit) -> Option<Timestamp>

Convert a timestamp to given time unit. Conversion from a timestamp with smaller unit to a larger unit will round the value to ceil (positive infinity). Return None if conversion causes overflow.

pub fn split(&self) -> (i64, u32)

Split a Timestamp into seconds part and nanoseconds part. Notice the seconds part of split result is always rounded down to floor.

fn from_splits(sec: i64, nsec: u32) -> Option<Self>

Creates a new Timestamp instance from seconds and nanoseconds parts. Returns None if overflow.

pub fn to_iso8601_string(&self) -> String

Format timestamp to ISO8601 string. If the timestamp exceeds what chrono timestamp can represent, this function simply print the timestamp unit and value in plain string.

pub fn to_local_string(&self) -> String

Format timestamp use system timezone.

pub fn to_timezone_aware_string(&self, tz: Option<&Timezone>) -> String

Format timestamp for given timezone. If tz==None, the server default timezone will used.

pub fn as_formatted_string( self, pattern: &str, timezone: Option<&Timezone>, ) -> Result<String>

Format timestamp for given format and timezone. If tz==None, the server default timezone will used.

pub fn to_chrono_datetime(&self) -> Option<NaiveDateTime>

pub fn to_chrono_datetime_with_timezone( &self, tz: Option<&Timezone>, ) -> Option<NaiveDateTime>

pub fn to_chrono_date(&self) -> Option<NaiveDate>

Convert timestamp to chrono date.

pub fn to_chrono_time(&self) -> Option<NaiveTime>

Convert timestamp to chrono time.

pub fn from_chrono_datetime(ndt: NaiveDateTime) -> Option<Self>

pub fn from_chrono_date(date: NaiveDate) -> Option<Self>

pub fn from_str_utc(s: &str) -> Result<Self>

Accepts a string in RFC3339 / ISO8601 standard format and some variants and converts it to a nanosecond precision timestamp. It no timezone specified in string, it cast to nanosecond epoch timestamp in UTC.

pub fn from_str(s: &str, timezone: Option<&Timezone>) -> Result<Self>

Accepts a string in RFC3339 / ISO8601 standard format and some variants and converts it to a nanosecond precision timestamp. This code is copied from arrow-datafusion with some bugfixes. Supported format:

2022-09-20T14:16:43.012345Z (Zulu timezone)

2022-09-20T14:16:43.012345+08:00 (Explicit offset)

2022-09-20T14:16:43.012345 (The given timezone, with T)

2022-09-20T14:16:43 (Zulu timezone, no fractional seconds, with T)

2022-09-20 14:16:43.012345Z (Zulu timezone, without T)

2022-09-20 14:16:43 (The given timezone, without T)

2022-09-20 14:16:43.012345 (The given timezone, without T)

pub fn negative(self) -> Self

impl Timestamp

pub const MIN_SECOND: Self

pub const MAX_SECOND: Self

pub const MIN_MILLISECOND: Self

pub const MAX_MILLISECOND: Self

pub const MIN_MICROSECOND: Self

pub const MAX_MICROSECOND: Self

pub const MIN_NANOSECOND: Self

pub const MAX_NANOSECOND: Self