common_time
Module timezone

Enums

Timezone

Statics

DEFAULT_TIMEZONE 🔒

System timezone in frontend/standalone, config by option default_timezone in toml, default value is UTC when default_timezone is not set.

Functions

get_timezone

If the tz=Some(timezone), return timezone directly, or return current system timezone.

parse_timezone

If the tz = Some("") || None || Some(Invalid timezone), return system timezone, or return parsed tz as timezone.

set_default_timezone

system_timezone_name

Return current system config timezone, default config is UTC