common_time::ttl

Enum TimeToLive

pub enum TimeToLive {
    Instant,
    Forever,
    Duration(Duration),
}

Time To Live

Variants

Instant

Instantly discard upon insert

Forever

Keep the data forever

Duration(Duration)

Duration to keep the data, this duration should be non-zero

Implementations

impl TimeToLive

pub fn from_humantime_or_str(s: &str) -> Result<Self, Error>

Parse a string that is either instant, forever, or a duration to TimeToLive

note that an empty string or a zero duration(a duration that spans no time) is treat as forever too

pub fn is_expired( &self, created_at: &Timestamp, now: &Timestamp, ) -> Result<bool>

Check if the TimeToLive is expired with the given created_at and now timestamp

pub fn is_instant(&self) -> bool

is instant variant

pub fn is_forever(&self) -> bool

Is the default value, which is Forever